requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.10
psycopg2-binary>=2.9.0 